Added "AI Images" Project
Added new Unity project that generates AI images at runtime. The scene "Main Scene" allows user to enter a prompt, and the AI image will be generated. Then, a new scene is automatically loaded, where the AI image is displayed in 2 formats - as a 2D image on a plane in front of the camera, and as a cubemap skybox texture.