New Unity Project with foundations for Text-to-Image Pipeline
Created a new Unity project, "Text-to-Image Pipeline", which will be where the main pipeline prototype is tested. Currently it can make an API call to Flux-360 model, which returns a 360-degree image. Plan to remove the "AI Images" project, as this new project will serve as the replacement for it soon.