Rendering 360-degree image in new scene
Added the functionality to take the image generated from Flux-360 model, and render it in a new scene in the Unity project. It is rendered as the skybox texture, and appears in the proper format, allowing the user to look around to move the camera and view the image as if inside of it.