Text Generation AI
Added call to text generation AI model - in this case, we are using GPT-2. It currently is returning an extended prompt; however, the quality of the prompt is poor, and is not working the way we desire it to. Basic functionality is in place, but there is much improvement needed.